[Bronze V] Title: 이상한 기호, Time: 176 ms, Memory: 17676 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -4,12 +4,12 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
 
-        int A = sc.nextInt();
-        int B = sc.nextInt();
+        long A = sc.nextInt();
+        long B = sc.nextInt();
 
         System.out.println(function(A ,B));
     }
-    public static long function(int A, int B) {
+    public static long function(long A, long B) {
         return (A + B) * (A - B);
 
     }
